requirements.txt
Flask>=2.1
Flask-SQLAlchemy>=3.0
argon2-cffi>=21.3.0
requests>=2.32.5
